refactor: extract OBS engine bridge from lib.rs to fix file size check
Move scene_change_commands and run_obs_engine_bridge into obs_bridge.rs
to bring lib.rs under the 1000-line CI limit.
